fix(vendor-finance): Extract duplicated logic
