+++ src/apps/vendor-finance/common/getBorrowerInfosQuery.ts
