--- src/apps/vendor-finance/graphql/constants.ts
@@ -1,2 +1,0 @@
-export const VENDOR_GRAPH_URL_ARBI = 'https://api.thegraph.com/subgraphs/name/0xtaiga/vendor-finance';
-export const VENDOR_GRAPH_URL_MAINNET = 'https://api.thegraph.com/subgraphs/name/0xtaiga/vendor-finance-mainnet';
